requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt.new
beautifulsoup4==4.12.2
    # via
    #   hn-rss-translator (pyproject.toml)
    #   deep-translator
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
deep-translator==1.11.4
    # via hn-rss-translator (pyproject.toml)
feedparser==6.0.10
    # via hn-rss-translator (pyproject.toml)
filelock==3.20.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
hf-xet==1.1.10
    # via huggingface-hub
huggingface-hub==0.35.3
    # via
    #   tokenizers
    #   transformers
idna==3.11
    # via requests
jinja2==3.1.6
    # via torch
lxml==4.9.3
    # via hn-rss-translator (pyproject.toml)
markupsafe==3.0.3
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.4
    # via transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.3.20
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
python-dotenv==1.0.0
    # via hn-rss-translator (pyproject.toml)
pyyaml==6.0.1
    # via
    #   hn-rss-translator (pyproject.toml)
    #   huggingface-hub
    #   transformers
regex==2025.10.23
    # via transformers
requests==2.31.0
    # via
    #   hn-rss-translator (pyproject.toml)
    #   deep-translator
    #   huggingface-hub
    #   transformers
safetensors==0.6.2
    # via transformers
setuptools==80.9.0
    # via torch
sgmllib3k==1.0.0
    # via feedparser
soupsieve==2.8
    # via beautifulsoup4
sympy==1.14.0
    # via torch
tokenizers==0.15.2
    # via transformers
torch==2.9.0
    # via hn-rss-translator (pyproject.toml)
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.35.2
    # via hn-rss-translator (pyproject.toml)
triton==3.5.0
    # via torch
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   torch
urllib3==2.5.0
    # via requests
